requirements.txt
beautifulsoup4==4.12.3
chromadb==0.5.5
openai==1.37.1
pandas==2.2.2
pdf2image==1.17.0
python-dotenv==1.0.1
Requests==2.32.3
rich==13.7.1
tqdm==4.65.0
